rat 10, auto mesh and more [BBPP40-371, BBPP40-384, BBPP40-385, BBPP40-386, BBPP40-392, BBPP40-394, BBPP40-395, BBPP40-396]
With this we introduce the new, production ready rat 10 circuit. At the moment it can run up to \~100 ms. After, it fails because some concentrations in steps go negative. This is not due to a "bug" per se but it is because the currents used to update the concentrations are too negative and strong. @shichkov suggests to tweak the initial parameters to solve this.

Since it is not a problem of the simulator per se, I would go ahead and merge this to use it as benchmark for future improvements

* move to github neurodamus

how to run rat10:

```
sbatch run_ratV10.slurm
```

**Auto-mesh** \[BBPP40-371, BBPP40-384, BBPP40-385, BBPP40-386, BBPP40-392, BBPP40-394, BBPP40-395\]

* remove os for pathing
* better config
* automatic mesh generation (cubic)
* throw error only if all the neurons have been dropped. Raise only a warning if a rank drops all its neurons \[BBPP40-396\]

Fixes:

* bf_m and ndam_m were not returning scaled points if we were asking for the points twice \[BBPP40-392\]
* \[BBPP40-394\]