improve ratV10 [BBPP40-397]
Neurodamus was failing with ratV10. I added a quickfix as suggested by @leite in \[BBPBGLIB-1068\]. Now the simulation dies at 27 ms because the currents from ndam are too strong and when converted to Na ions it makes the concentrations to go negative. This may be further investigated by @sfarina .

**How to run ratV10**

Once youa re on the correct branch you just need to:

```
sbatch run_ratV10.slurm
```