Dismember partially init inputs. Fix: [BBPP40-441]
In the process of solving BBPP40-434 we need to have a more flexible way of assigning ndam, steps, and bf vars. Atm everything is done in init inputs. We need to extract and split it in 3 methods. For the moment this should be equivalent (in performance) to what was done before. Maybe an O(1) slower. The main advantage is that, if this works, I can proceed to use these methods in other points of the code (not when init_inputs is called only). Plus, the code is clearer and more modular in this way.

Results should remain exactly the same

I know init_inputs is a mess but it should disappear in a following MR (It will happen soon).

Fix: \[BBPP40-441\]