init_inputs -> check_inputs and init metab [BBPP40-444, BBPP40-445]
This is a merge of 2 MRs.

**check_inputs**

Now that init_inputs has been largely dismembered into chunks and moved out of the function the only thing that remains (for this function) is to turn it in what it really should be: a function to check the metab inputs (vm) and params before the advance call.

Preparatory for BBPP40-434

Fix: BBPP40-444

**Init metab**

Param may also turn in a standard list. I am not sold on that detail. For sure it is better to have all the checks in the same function

Metab never needed an init before the simulations because it was always after the other simulator steps in the pipeline. In other words its init was called at the beginning of a metab time step every time based on the results of the other simulators.

Now we want it to be a first class citizen and compute "in parallel" with the other simulators (and merge discording results after).

This means that we need to initialize it. With what? With the results of the other simulators. Thus, it is still dependent on the other simulators but only for the initial conditions.

This MR is a transitioning one. We still have the previous integration scheme with the new setting of the initial conditions. Thus, what I did here should not affect the results except for an additional sync at the beginning of the simulation (rewritten before being used). Performance should stay the same or go down by an O(1).

Fix: \[BBPP40-445\] Preparatory for \[BBPP40-434\]

**Notes**

steps_vars, bloodflow_vars, ndam_vars are still there only for reporting purposes. They will be erased in a future MR pretty soon I hope.

All in all this MR should not change results. We may see it be slower by an O(1) as we are adding an additional initialization. It is useless for now. It will become useful in a future MR that addresses \[BBPP40-434\]