rework metabolism [BBPP40-434, BBPP40-466, BBPP40-455, BBPP40-440, BBPP40-427, BBPP40-470]
This MR contains major reworks for:

* the config object
* the metabolism object
* the reporter object
* the connection object

Dan wanted to have a more scientific way of merging conflicting values from neurodamus and metabolism. This generated a cascade of events that made a lot of stuff obsolete or clunky to work with. For this reason I went forward and reworked a lot of stuff behind the scenes.

The initial aim was to get rid of all (-most?) the magic numbers springled here and there by Polina in the glue code. Now there are only 3 ways to sync data from a source simulation to a destination simulator:

* add: the values from the source are added to the destination
* set: the values from the source replace the values in destination
* merge: source and destination values are combined with this formula: V_source + V_destination - V_n-1 = V_n. Qualitatively here we sum the contributions from source and destination to get the final, merged value. This final value is then set for source and destination. Notice that swapping source and destination produces the same result. This is the only scheme that is flipping-invariant

Why did this generate so much change? Because a lot of stuff in the config file became unnecessary or complicated to maintain in the current form. Since we need also to merge msr_config inside simulation_config and make it sonata compliant I went forward and did both things.

Fix: BBPP40-434, BBPP40-466, BBPP40-455, BBPP40-440, BBPP40-427, BBPP40-470

Small changes:

* improved documentation
* updated postproc
* more pytest logging
* make simulations die correctly when an error is raised with mpi4py

TODO:

* [x] rat V10 move 400 -\> 4000 ndts for bf and metab
* [x] remove debug code: `# @utils.apply_abort_on_exception_decorator`
* [x] documentation for the new config file

Postponed to a future MR:

* BBPP40-471