Eval Python expr from JSON in a restricted env
Replace calls to `eval(expr)` where `expr` are Python expressions specified in the JSON configuration to modify the values transfered from a model to another. This is trivially sensible to code injection attack.
Instead, rely on the `simpleeval` Python package that allows one to define a restricted environment where a Python expression is evaluated (where the symbol `shutil.rmtree` is not available for instance)

Only the following symbols are available within the Python execution environment:
* variables:
  * `config`: the `MsrConfig` instance of the simulation
  * `vals`: the data routing from a model to another
* modules:
  * `np`: the NumPy module
  * `math`: from the standard library
* computational Python builtins: abs, min, max, pow, round, sum