Set can set also the source if present [BBPBGLIB-1159, BBPP40-485]
The set action at the moment can only set values from the source simulator to the destination. This change makes so that it can set, if the source_set variables are present, also the values in the source.

In addition, we also add the possibility for variable conversion.

For example, we can set the ADPI in metabolism and neurodamus to be a function of the ATPI in neurodamus and be marked in both neurodamus and metabolism every time there is a metabolism sync

Fix \[ BBPBGLIB-1159, BBPP40-485\]